Moved Google ESP code to a separate file.
Moved Identity type and helpers to a separate file.
Tidied up configuration passing to dependencies.
Added a separate file for code to handle the init callback.
Fixed a bug where the promise handler was throwing a different error than expected.
Fixed a few tests which seemed wrong.
Refactored the logic involved in setting an identity.